added unit range detector

+++ Assets/ScriptableObjs/UnitTypes/UnitSO.cs
@@ -11,6 +11,6 @@ public class UnitSO : ScriptableObject
     public Sprite unitIcon;
     public UnitType unitType;
     public int MaxHP;
-    public int DetectRange;
-    public int AttackRange;
+    public float DetectRange;
+    public float AttackRange;
 }
